+++ src/utils_GUI/event_functions.jl
@@ -150,7 +150,6 @@ function define_event_functions(gui::GUI)
                             clear_selection(gui; clear_topo=false)
                         end
                         pick_component!(gui; pick_results_component=true)
-                        return Consume(true)
                     end
                     return Consume(false)
                 end
@@ -302,7 +301,7 @@ function define_event_functions(gui::GUI)
             selection[:visible] = false
             selection[:pinned] = false
         end
-        clear_selection(gui; clear_results=true)
+        clear_selection(gui; clear_topo=false)
         update_legend!(gui)
         return Consume(false)
     end
@@ -359,13 +358,13 @@ function define_event_functions(gui::GUI)
                 end
             elseif axes_str == "All"
                 model = get_model(gui)
-                for dict ∈ collect(keys(object_dictionary(model)))
-                    container = model[dict]
+                for sym ∈ get_JuMP_names(gui)
+                    container = model[sym]
                     if isempty(container)
                         continue
                     end
                     if typeof(container) <: JuMP.Containers.DenseAxisArray
-                        axis_types = nameof.([eltype(a) for a ∈ JuMP.axes(model[dict])])
+                        axis_types = nameof.([eltype(a) for a ∈ JuMP.axes(model[sym])])
                     elseif typeof(container) <: SparseVars
                         axis_types = collect(nameof.(typeof.(first(keys(container.data)))))
                     end