+++ src/utils_GUI/info_axis_utils.jl
@@ -4,18 +4,18 @@
 Based on `element` update the text in info box.
 """
 function update_info_box!(gui::GUI, element; indent::Int64=0)
-    infoBox = get_ax(gui, :info).scene.plots[1][1]
+    info_box = get_ax(gui, :info).scene.plots[1][1]
     if isnothing(element)
         investment_overview = get_var(gui, :investment_overview)
         if isempty(investment_overview)
-            infoBox[] = get_var(gui, :default_text)
+            info_box[] = get_var(gui, :default_text)
         else
-            infoBox[] = investment_overview
+            info_box[] = investment_overview
         end
         return nothing
     end
     if indent == 0
-        infoBox[] = "$element ($(typeof(element)))\n"
+        info_box[] = "$element ($(typeof(element)))\n"
     end
     indent += 1
     indent_str::String = "  "^indent
@@ -24,24 +24,31 @@ function update_info_box!(gui::GUI, element; indent::Int64=0)
     if isa(element, Vector)
         for (i, field1) ∈ enumerate(element)
             if is_iterable(field1)
-                infoBox[] *= indent_str * "$i: $(typeof(field1)):\n"
+                info_box[] *= indent_str * "$i: $(typeof(field1)):\n"
                 update_info_box!(gui, field1; indent)
             else
-                infoBox[] *= indent_str * "$i: $(typeof(field1))\n"
+                info_box[] *= indent_str * "$i: $(typeof(field1))\n"
             end
         end
     elseif isa(element, Dict)
         for field1 ∈ keys(element)
-            infoBox[] *= indent_str * "$field1 => $(element[field1])\n"
+            info_box[] *= indent_str * "$field1 => $(element[field1])\n"
         end
     else
         for field1 ∈ fieldnames(typeof(element))
             value1 = getfield(element, field1)
             if is_iterable(value1)
-                infoBox[] *= indent_str * "$(field1) ($(typeof(value1))):\n"
+                info_box[] *= indent_str * "$(field1) ($(typeof(value1))):\n"
                 update_info_box!(gui, value1; indent)
             else
-                infoBox[] *= indent_str * "$(field1): $value1\n"
+                if isa(value1, OperationalProfile) &&
+                    !isa(value1, FixedProfile) &&
+                    length(value1.vals) > 168
+                    # Truncate large vectors
+                    info_box[] *= indent_str * "$(field1): $(typeof(value1))\n"
+                else
+                    info_box[] *= indent_str * "$(field1): $value1\n"
+                end
             end
         end
     end